LayoutTests/fast/scrolling/mac/slow-scrolling-overflow-expected.txt


Test that scroll over scroller with background-attachment:fixed content works.
PASS overflowScrollEventCount > 0 is true
PASS successfullyParsed is true

TEST COMPLETE
